+++ src/Source.h
@@ -14,16 +14,16 @@ MatrixXf Gaussian2D(int Nx, int Ny, double dx, double dy, double sigma_x, double
 
 struct Coor2D
 {
-    uint16_t x, y;
+    int x, y;
     Coor2D() : x(0), y(0) {}
-    Coor2D(uint16_t x, uint16_t y) : x(x), y(y) {}
-    uint16_t X() {
+    Coor2D(int x, int y) : x(x), y(y) {}
+    int X() {
         return x;
     }
-    uint16_t Y() {
+    int Y() {
         return y;
     }
-    const Coor2D& operator+=(uint16_t N) {
+    const Coor2D& operator+=(int N) {
         x += N;
         y += N;
         return *this;
@@ -33,23 +33,23 @@ struct Coor2D
 class PointSource2D
 {
     // N: the number of point source
-    uint16_t N = 0;
+    int N = 0;
     std::vector<Coor2D> source_coor;
     std::vector<VectorXf> source_fn;
 public:
     PointSource2D();
-    PointSource2D(uint16_t coor_x, uint16_t coor_y);
-    PointSource2D(uint16_t coor_x, uint16_t coor_y, const VectorXf& fn);
+    PointSource2D(int coor_x, int coor_y);
+    PointSource2D(int coor_x, int coor_y, const VectorXf& fn);
     PointSource2D(const std::vector<Coor2D>& coor);
     PointSource2D(const std::vector<Coor2D>& coor, const VectorXf& fn);
     PointSource2D(const std::vector<Coor2D>& coor, const std::vector<VectorXf>& fn);
 
-    void UpdatePML(uint16_t N_pml);
+    void UpdatePML(int N_pml);
 
-    uint16_t GetSourceNum();
-    Coor2D GetCoor(uint16_t source_idx);
-    VectorXf GetFn(uint16_t source_idx);
-    double GetValue(uint16_t source_idx, uint16_t time_idx);
+    int GetSourceNum();
+    Coor2D GetCoor(int source_idx);
+    VectorXf GetFn(int source_idx);
+    double GetValue(int source_idx, int time_idx);
     void PrintInfo();
 };
